refactored Services in order to applied Dependency Inversion Principle by adding a layer of abstraction
